Change WeaponSlots position to int rather than float
(Incomplete Commit)

+++ Source/Engine/models.h
@@ -19,7 +19,7 @@
 
 typedef struct WeaponSlot {
 	string name;            ///< name of the slot
-	double x,y;             ///< The relative X and Y offsets of this weapon.
+	int x,y;             ///< The relative X and Y offsets of this weapon.
 	double angle;           ///< angle the weapon will be mounted
 	double motionAngle;     ///< should be either 0 (meaning no turrets allowed) or a number > 0 and <= 360
 	Weapon* content;         ///< name of the weapon it contains (or "" for empty)
